+++ brokerapi/brokers/pubsub/broker.go
@@ -21,6 +21,7 @@ import (
 	googlepubsub "cloud.google.com/go/pubsub"
 	"encoding/json"
 	"gcp-service-broker/brokerapi/brokers/models"
+	"gcp-service-broker/brokerapi/brokers/name_generator"
 	"golang.org/x/net/context"
 	"net/http"
 
@@ -48,8 +49,15 @@ func (b *PubSubBroker) Provision(instanceId string, details models.ProvisionDeta
 
 	var err error
 	var params map[string]string
-	if err := json.Unmarshal(details.RawParameters, &params); err != nil {
+	if len(details.RawParameters) == 0 {
+		params = map[string]string{}
+	} else if err := json.Unmarshal(details.RawParameters, &params); err != nil {
 		return models.ServiceInstanceDetails{}, fmt.Errorf("Error unmarshalling provision details: %s", err)
+	}
+
+	// Ensure there is a name for this topic
+	if _, ok := params["topic_name"]; !ok {
+		params["topic_name"] = name_generator.Basic.InstanceName()
 	}
 
 	ctx := context.Background()