+++ integration_tests/integration_test.go
@@ -1,13 +1,14 @@
 package integration_tests
 
 import (
-	"encoding/json"
 	. "gcp-service-broker/brokerapi/brokers"
 	"golang.org/x/net/context"
 
 	"gcp-service-broker/brokerapi/brokers"
 	"gcp-service-broker/brokerapi/brokers/models"
+	"gcp-service-broker/brokerapi/brokers/name_generator"
 	"gcp-service-broker/db_service"
+	"gcp-service-broker/fakes"
 	"net/http"
 	"os"
 
@@ -26,14 +27,13 @@ import (
 const timeout = 60
 
 type genericService struct {
-	serviceId          string
-	planId             string
-	bindingId          string
-	rawProvisionParams json.RawMessage
-	rawBindingParams   map[string]interface{}
-	instanceId         string
-	serviceExistsFn    func(bool) bool
-	cleanupFn          func()
+	serviceId        string
+	planId           string
+	bindingId        string
+	rawBindingParams map[string]interface{}
+	instanceId       string
+	serviceExistsFn  func(bool) bool
+	cleanupFn        func()
 }
 
 func testGenericService(gcpBroker *GCPAsyncServiceBroker, params *genericService) {
@@ -45,9 +45,8 @@ func testGenericService(gcpBroker *GCPAsyncServiceBroker, params *genericService
 	// Provision
 	//
 	provisionDetails := models.ProvisionDetails{
-		ServiceID:     params.serviceId,
-		PlanID:        params.planId,
-		RawParameters: params.rawProvisionParams,
+		ServiceID: params.serviceId,
+		PlanID:    params.planId,
 	}
 
 	_, err := gcpBroker.Provision(params.instanceId, provisionDetails, true)
@@ -125,6 +124,7 @@ var _ = Describe("LiveIntegrationTests", func() {
 		logger              lager.Logger
 		serviceNameToId     map[string]string = make(map[string]string)
 		serviceNameToPlanId map[string]string = make(map[string]string)
+		instance_name       string            = "pcf_sb_2_1479851465224535828"
 	)
 
 	BeforeEach(func() {
@@ -270,6 +270,8 @@ var _ = Describe("LiveIntegrationTests", func() {
 			}
 		}`)
 
+		name_generator.Basic = &fakes.StaticNameGenerator{Val: instance_name}
+
 		gcpBroker, err = brokers.New(logger)
 		if err != nil {
 			logger.Error("error", err)
@@ -316,23 +318,21 @@ var _ = Describe("LiveIntegrationTests", func() {
 			service, err := googlebigquery.New(gcpBroker.GCPClient)
 			Expect(err).NotTo(HaveOccurred())
 
-			datasetName := "integration_test_dataset"
 			params := &genericService{
-				serviceId:          serviceNameToId[brokers.BigqueryName],
-				planId:             serviceNameToPlanId[brokers.BigqueryName],
-				bindingId:          "integration_test_bind",
-				instanceId:         datasetName,
-				rawProvisionParams: []byte("{\"name\": \"" + datasetName + "\"}"),
+				serviceId:  serviceNameToId[brokers.BigqueryName],
+				planId:     serviceNameToPlanId[brokers.BigqueryName],
+				bindingId:  "integration_test_bind",
+				instanceId: "integration_test_dataset",
 				rawBindingParams: map[string]interface{}{
 					"role": "bigquery.admin",
 				},
-				serviceExistsFn: func(bool) bool {
-					_, err = service.Datasets.Get(gcpBroker.RootGCPCredentials.ProjectId, datasetName).Do()
+				serviceExistsFn: func(expected bool) bool {
+					_, err = service.Datasets.Get(gcpBroker.RootGCPCredentials.ProjectId, instance_name).Do()
 
 					return err == nil
 				},
 				cleanupFn: func() {
-					err := service.Datasets.Delete(gcpBroker.RootGCPCredentials.ProjectId, datasetName).Do()
+					err := service.Datasets.Delete(gcpBroker.RootGCPCredentials.ProjectId, instance_name).Do()
 					Expect(err).NotTo(HaveOccurred())
 				},
 			}
@@ -457,24 +457,22 @@ var _ = Describe("LiveIntegrationTests", func() {
 			service, err := googlestorage.NewClient(context.Background(), option.WithUserAgent(models.CustomUserAgent))
 			Expect(err).NotTo(HaveOccurred())
 
-			bucketName := "integration_test_bucket"
 			params := &genericService{
-				serviceId:          serviceNameToId[brokers.StorageName],
-				planId:             serviceNameToPlanId[brokers.StorageName],
-				instanceId:         bucketName,
-				bindingId:          "integration_test_bucket_binding",
-				rawProvisionParams: []byte("{\"name\": \"" + bucketName + "\"}"),
+				serviceId:  serviceNameToId[brokers.StorageName],
+				planId:     serviceNameToPlanId[brokers.StorageName],
+				instanceId: "integration_test_bucket",
+				bindingId:  "integration_test_bucket_binding",
 				rawBindingParams: map[string]interface{}{
 					"role": "storage.admin",
 				},
 				serviceExistsFn: func(bool) bool {
-					bucket := service.Bucket(bucketName)
+					bucket := service.Bucket(instance_name)
 					_, err = bucket.Attrs(context.Background())
 
 					return err == nil
 				},
 				cleanupFn: func() {
-					bucket := service.Bucket(bucketName)
+					bucket := service.Bucket(instance_name)
 					bucket.Delete(context.Background())
 				},
 			}
@@ -488,22 +486,19 @@ var _ = Describe("LiveIntegrationTests", func() {
 			service, err := googlepubsub.NewClient(context.Background(), gcpBroker.RootGCPCredentials.ProjectId, option.WithUserAgent(models.CustomUserAgent))
 			Expect(err).NotTo(HaveOccurred())
 
-			topicName := "integration_test_topic"
-			topic := service.Topic(topicName)
+			topic := service.Topic(instance_name)
 
 			params := &genericService{
-				serviceId:          serviceNameToId[brokers.PubsubName],
-				planId:             serviceNameToPlanId[brokers.PubsubName],
-				instanceId:         topicName,
-				bindingId:          "integration_test_topic_bindingId",
-				rawProvisionParams: []byte("{\"topic_name\": \"" + topicName + "\"}"),
+				serviceId:  serviceNameToId[brokers.PubsubName],
+				planId:     serviceNameToPlanId[brokers.PubsubName],
+				instanceId: "integration_test_topic",
+				bindingId:  "integration_test_topic_bindingId",
 				rawBindingParams: map[string]interface{}{
 					"role": "pubsub.admin",
 				},
 				serviceExistsFn: func(bool) bool {
 					exists, err := topic.Exists(context.Background())
-					Expect(err).NotTo(HaveOccurred())
-					return exists
+					return exists && err == nil
 				},
 				cleanupFn: func() {
 					err := topic.Delete(context.Background())